# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Permy \u2014 Building Permit & Construction Intelligence API",
    "description": "Normalized permits, contractor activity, property timelines, and ZIP development signals \u2014 ranked, sourced, and ready for your app. Give Permy an address, city, ZIP, contractor, or trade and get back clean, machine-readable municipal intelligence with source links and confidence scores.",
    "termsOfService": "https://docs.permy.dev/terms",
    "contact": {
      "name": "Permy",
      "url": "https://permy.dev/",
      "email": "hi@permy.dev"
    },
    "license": {
      "name": "Apache-2.0 (code); data per upstream license",
      "url": "https://docs.permy.dev/legal"
    },
    "version": "0.1.0",
    "x-logo": {
      "url": "https://permy.dev/assets/brand-mark.png"
    }
  },
  "servers": [
    {
      "url": "https://permy.p.rapidapi.com",
      "description": "RapidAPI production"
    },
    {
      "url": "https://api.permy.dev",
      "description": "Direct site"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local dev"
    }
  ],
  "paths": {
    "/v1/permits/search": {
      "get": {
        "tags": [
          "permits"
        ],
        "summary": "Search normalized permits",
        "operationId": "search_permits_v1_permits_search_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "zip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zip"
            }
          },
          {
            "name": "trade",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "roofing|solar|hvac|plumbing|electrical|building|general|demolition|other|unknown",
              "title": "Trade"
            },
            "description": "roofing|solar|hvac|plumbing|electrical|building|general|demolition|other|unknown"
          },
          {
            "name": "permit_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Permit Type"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "applied|issued|active|final|expired|cancelled|withdrawn|unknown",
              "title": "Status"
            },
            "description": "applied|issued|active|final|expired|cancelled|withdrawn|unknown"
          },
          {
            "name": "contractor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Contractor name (substring)",
              "title": "Contractor"
            },
            "description": "Contractor name (substring)"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text over description",
              "title": "Keyword"
            },
            "description": "Free-text over description"
          },
          {
            "name": "min_valuation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Valuation"
            }
          },
          {
            "name": "max_valuation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Valuation"
            }
          },
          {
            "name": "issued_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issued After"
            }
          },
          {
            "name": "issued_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issued Before"
            }
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "west,south,east,north \u2014 geo bounding box",
              "title": "Bbox"
            },
            "description": "west,south,east,north \u2014 geo bounding box"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(issued_date|valuation_usd|lead_score)$",
              "default": "issued_date",
              "title": "Sort"
            }
          },
          {
            "name": "sort_dir",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "desc",
              "title": "Sort Dir"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermitsSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/permits/{permit_id}": {
      "get": {
        "tags": [
          "permits"
        ],
        "summary": "Full permit detail with enrichment + source links",
        "operationId": "get_permit_v1_permits__permit_id__get",
        "parameters": [
          {
            "name": "permit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Permit Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permit"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/properties/resolve": {
      "get": {
        "tags": [
          "permits"
        ],
        "summary": "Normalize an address \u2192 jurisdiction, city/county, parcel hints, coverage status",
        "operationId": "resolve_property_v1_properties_resolve_get",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Full street address, e.g. '10912 Mystic Timber Dr, Austin, TX 78754'",
              "title": "Address"
            },
            "description": "Full street address, e.g. '10912 Mystic Timber Dr, Austin, TX 78754'"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Property"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/properties/{property_id}/timeline": {
      "get": {
        "tags": [
          "permits"
        ],
        "summary": "Full permit history for an address",
        "operationId": "property_timeline_v1_properties__property_id__timeline_get",
        "parameters": [
          {
            "name": "property_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Property Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyTimeline"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/v1/sample/permits/{permit_id}": {
      "get": {
        "tags": [
          "sample"
        ],
        "summary": "[Sample, no key] Full permit detail",
        "operationId": "sample_get_permit_v1_sample_permits__permit_id__get",
        "parameters": [
          {
            "name": "permit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Permit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permit"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OwnerRef": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OwnerRef"
      },
      "Permit": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Permy opaque id (stringified bigint)"
          },
          "canonical_uid": {
            "type": "string",
            "title": "Canonical Uid",
            "description": "Stable hash(jurisdiction + source_permit_id)"
          },
          "jurisdiction_slug": {
            "type": "string",
            "title": "Jurisdiction Slug"
          },
          "source_permit_id": {
            "type": "string",
            "title": "Source Permit Id"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "first_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen At"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          },
          "last_checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Checked At"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "permit_type_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permit Type Raw"
          },
          "permit_type_normalized": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permit Type Normalized"
          },
          "work_class": {
            "type": "string",
            "enum": [
              "new_construction",
              "alteration",
              "addition",
              "remodel",
              "repair",
              "demolition",
              "other",
              "unknown"
            ],
            "title": "Work Class",
            "default": "unknown"
          },
          "trade_category": {
            "type": "string",
            "enum": [
              "roofing",
              "solar",
              "hvac",
              "plumbing",
              "electrical",
              "building",
              "general",
              "demolition",
              "other",
              "unknown"
            ],
            "title": "Trade Category",
            "default": "unknown"
          },
          "is_new_construction": {
            "type": "boolean",
            "title": "Is New Construction",
            "default": false
          },
          "is_alteration": {
            "type": "boolean",
            "title": "Is Alteration",
            "default": false
          },
          "is_demolition": {
            "type": "boolean",
            "title": "Is Demolition",
            "default": false
          },
          "valuation_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valuation Usd",
            "description": "Declared job value in USD; null when city doesn't publish"
          },
          "housing_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Housing Units"
          },
          "new_add_sqft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Add Sqft"
          },
          "dates": {
            "$ref": "#/components/schemas/PermitDates"
          },
          "current_status": {
            "type": "string",
            "enum": [
              "applied",
              "issued",
              "active",
              "final",
              "expired",
              "cancelled",
              "withdrawn",
              "unknown"
            ],
            "title": "Current Status",
            "default": "unknown"
          },
          "status_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Raw"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "description_enriched": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description Enriched"
          },
          "contractor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContractorRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "parcel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parcel Id"
          },
          "enrichment": {
            "$ref": "#/components/schemas/Enrichment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "canonical_uid",
          "jurisdiction_slug",
          "source_permit_id",
          "first_seen_at",
          "last_seen_at",
          "last_checked_at",
          "address"
        ],
        "title": "Permit"
      },
...
      "PermitsSearchResponse": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "permits": {
            "items": {
              "$ref": "#/components/schemas/Permit"
            },
            "type": "array",
            "title": "Permits"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "permits"
        ],
        "title": "PermitsSearchResponse"
      },
      "Property": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "canonical_uid": {
            "type": "string",
            "title": "Canonical Uid"
          },
          "full_address": {
            "type": "string",
            "title": "Full Address"
          },
          "street": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Street"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "zip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "geocode_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geocode Confidence"
          },
          "jurisdiction_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction Slug"
          },
          "parcel_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parcel Id"
          },
          "year_built": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Built"
          },
          "sqft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sqft"
          },
          "permit_count": {
            "type": "integer",
            "title": "Permit Count",
            "default": 0
          },
          "last_permit_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Permit Date"
          },
          "coverage_status": {
            "type": "string",
            "enum": [
              "covered",
              "partial",
              "no_feed"
            ],
            "title": "Coverage Status",
            "default": "covered"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "canonical_uid",
          "full_address"
        ],
        "title": "Property"
      },
      "PropertyTimeline": {
        "properties": {
          "property": {
            "$ref": "#/components/schemas/Property"
          },
          "permits": {
            "items": {
              "$ref": "#/components/schemas/PermitTimelineEntry"
            },
            "type": "array",
            "title": "Permits"
          },
          "total_permits": {
            "type": "integer",
            "title": "Total Permits"
          },
          "total_valuation_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Valuation Usd"
          },
          "last_activity": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity"
          },
          "unpermitted_work_flag": {
            "type": "boolean",
            "title": "Unpermitted Work Flag",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "property",
          "permits",
          "total_permits"
        ],
        "title": "PropertyTimeline"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Key",
        "description": "Your RapidAPI app key (sent automatically by the gateway). Direct-site callers may use X-API-Key or Authorization: Bearer instead."
      }
    }
  },
...
}